requirement.txt
numpy
scipy
scikit-learn
pytorch-gpu
tqdm
ipdb
wandb
k-means-constrained
